Fixup circular dependence
